automated_price_tracking/requirements.txt
streamlit
firecrawl-py>=4.12.0
pydantic
psycopg2-binary
python-dotenv
sqlalchemy==2.0.35
pandas
plotly
aiohttp
